+++ src/DTO/Config/GlobalConfig.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace DTO\Config;
+
+readonly class GlobalConfig
+{
+    public function __construct(
+        public int $requestSizeMax,
+        public ?PreforkConfig $prefork,
+        public array $hosts,
+    ) {
+    }
+}