+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning v2.0.0](http://semver.org/).
 
-## [0.3.0] - 2026-01-27
+## [0.3.0] - 2026-01-29
 ### Added
+- Global configuration using XML file
 - Prefork server model
 - Signal processing
 - HTTPS support